+++ examples/resources/meshstack_tenant_v4/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  id = "00000000-0000-0000-0000-000000000000" # tenant uuid
-  to = meshstack_tenant_v4.example
-}